Lab2/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform.
# Manually managing azure-functions-worker may cause unexpected issues.

azure-functions
pymongo
redis
python-dotenv
requests
